requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.6.3
botocore==1.9.3
bs4==0.0.1
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
click==6.7
cssselect==1.0.3
cycler==0.10.0
docutils==0.14
gensim==3.4.0
idna==2.6
jmespath==0.9.3
lda==1.0.5
lxml==4.1.1
matplotlib==2.1.2
nltk==3.2.5
numpy==1.14.1
pbr==3.1.1
preprocessor==1.1.3
pyparsing==2.2.0
pyquery==1.4.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
smart-open==1.5.6
toolz==0.9.0
tweet-preprocessor==0.5.0
urllib3==1.22
